fix(compiler-wasm): add shim-free browser-only ./web subpath export
Split browser-safe facade (createCompilerFromWasmModule + build helpers +
shared types) into src/web.ts, which has no pkg-node import in its graph.
tsup builds both src/index.ts and src/web.ts; dist/web.mjs gets no
fileURLToPath/init_esm_shims chunk. The ./web subpath export is added to
the package exports map. Main-entry public API is unchanged — index.ts
re-exports everything from ./web and keeps loadWasm/createCompiler/
createCompilerFromSnapshot for Node consumers. Regression test in
__tests__/web-entry.test.ts asserts shim strings are absent from dist/web.mjs.
